A/b test the book page
